cleanup: separate the code into modules
Separating modules will help to unit test the code and structures
the code, which makes navigation much easier.
