+++ firmware/buttons/CMakeLists.txt
@@ -1,0 +1,13 @@
+set(modulename "buttons")
+set(SOURCES 
+        buttons.cpp
+)
+add_library(${modulename} ${SOURCES})
+target_include_directories(${modulename} PUBLIC include)
+target_link_libraries(${modulename}
+    pico_stdlib 
+    tinyusb_board
+    tinyusb_device
+    usb
+    leds
+)
