+++ firmware/usb/include/hid.hpp
@@ -1,0 +1,5 @@
+#pragma once
+
+#include "buttons.hpp"
+
+void hid_task(const Buttons& buttons);
