+++ firmware/usb/include/tud.hpp
@@ -1,0 +1,5 @@
+#pragma once
+
+#include "bsp/board_api.h"
+
+void initialize_tud();
