+++ firmware/usb/include/usb_descriptors.h
@@ -1,0 +1,8 @@
+#ifndef USB_DESCRIPTORS_H_
+#define USB_DESCRIPTORS_H_
+
+enum {
+    REPORT_ID_KEYBOARD = 1,
+};
+
+#endif /* USB_DESCRIPTORS_H_ */
